refactor: adjust email service into it's own package
