--- src/main/java/com/jpcchaves/adotar/application/service/usecases/EmailService.java
@@ -1,14 +1,0 @@
-package com.jpcchaves.adotar.application.service.usecases;
-
-import com.jpcchaves.adotar.application.dto.ApiMessageResponseDto;
-import com.jpcchaves.adotar.application.dto.email.ContactEmailDto;
-import com.jpcchaves.adotar.domain.model.PasswordResetToken;
-import jakarta.mail.MessagingException;
-
-public interface EmailService {
-  void sendResetPasswordRequest(PasswordResetToken passwordResetRequestDto)
-      throws MessagingException;
-
-  ApiMessageResponseDto sendContactMessage(ContactEmailDto contactEmailDto)
-      throws MessagingException;
-}
