+++ src/main/java/com/jpcchaves/adotar/presentation/controller/v1/ContactUsController.java
@@ -2,7 +2,7 @@ package com.jpcchaves.adotar.presentation.controller.v1;
 
 import com.jpcchaves.adotar.application.dto.ApiMessageResponseDto;
 import com.jpcchaves.adotar.application.dto.email.ContactEmailDto;
-import com.jpcchaves.adotar.application.service.usecases.EmailService;
+import com.jpcchaves.adotar.application.service.mail.contracts.EmailService;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.media.Content;
 import io.swagger.v3.oas.annotations.media.Schema;
@@ -19,35 +19,35 @@ import org.springframework.web.bind.annotation.*;
 @CrossOrigin(origins = "*")
 @Tag(name = "Contact Us-Controller")
 public class ContactUsController {
-  private final EmailService emailService;
+    private final EmailService emailService;
 
-  public ContactUsController(EmailService emailService) {
-    this.emailService = emailService;
-  }
+    public ContactUsController(EmailService emailService) {
+        this.emailService = emailService;
+    }
 
-  @PostMapping
-  @Operation(
-      summary = "Sends a contact email",
-      description = "Sends a contact email to the Adote.ME support",
-      responses = {
-        @ApiResponse(
-            description = "Success",
-            responseCode = "200",
-            content =
-                @Content(
-                    schema =
-                        @Schema(implementation = ApiMessageResponseDto.class))),
-        @ApiResponse(
-            description = "Bad Request",
-            responseCode = "400",
-            content = @Content),
-        @ApiResponse(
-            description = "Internal Error",
-            responseCode = "500",
-            content = @Content),
-      })
-  public ResponseEntity<ApiMessageResponseDto> sendContactUsMessage(
-      @RequestBody ContactEmailDto contactEmailDto) throws MessagingException {
-    return ResponseEntity.ok(emailService.sendContactMessage(contactEmailDto));
-  }
+    @PostMapping
+    @Operation(
+            summary = "Sends a contact email",
+            description = "Sends a contact email to the Adote.ME support",
+            responses = {
+                    @ApiResponse(
+                            description = "Success",
+                            responseCode = "200",
+                            content =
+                            @Content(
+                                    schema =
+                                    @Schema(implementation = ApiMessageResponseDto.class))),
+                    @ApiResponse(
+                            description = "Bad Request",
+                            responseCode = "400",
+                            content = @Content),
+                    @ApiResponse(
+                            description = "Internal Error",
+                            responseCode = "500",
+                            content = @Content),
+            })
+    public ResponseEntity<ApiMessageResponseDto> sendContactUsMessage(
+            @RequestBody ContactEmailDto contactEmailDto) throws MessagingException {
+        return ResponseEntity.ok(emailService.sendContactMessage(contactEmailDto));
+    }
 }
